requirements.txt
scikit-learn==0.22.1
pandas==1.0.3
numpy==1.18.1
joblib==0.14.1
torch==1.7.1
nltk==3.2.5
spacy==2.1.8
fastai==1.0.51
gensim==3.8.3
sklearn-pandas==1.8.0
psutil==5.6.7
Flask==1.0.2
